requirements.txt
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
boto3==1.33.0 